refac(rfr): move rec module to streamed
The `rec` module in the `rfr` crate is a bit of a relic from when the
streamed recording format was the only format. It is now something of a
misnomer. Additionally, the `AbsTimestamp` type got left in that module,
despite being used in the chunked recording format as well.

This change moves everything out of the `rec` module into a new
`streamed` module which matches the layout of the `chunked` module
(separate `read` and `write` modules inside).

An exception is `AbsTimestamp`, which moves to the `common` module, but
is re-exported at the root of the `rfr` crate. The plan is to make the
`common` module private and re-export any public types in the same way.

One further exception is the `WinTimestamp` struct that was in the `rec`
module. This type is only used in the `rfr-viz` crate, so it was moved
there and made crate internal.
